refactor: move error_chain_fmt to a crate-level errors module
